+++ modules/mod-playerbots/src/AI/Base/Actions/TravelAction.cpp
@@ -1,6 +1,7 @@
 /*
- * Copyright (C) 2016+ AzerothCore <www.azerothcore.org>, released under GNU AGPL v3 license, you may redistribute it
- * and/or modify it under version 3 of the License, or (at your option), any later version.
+ * This file is part of the mod-playerbots module for AzerothCore. See AUTHORS file for Copyright
+ * information; released under GNU GPL v2 license, redistribute/modify under version 2 of the License,
+ * or (at your option) any later version.
  */
 
 #include "TravelAction.h"
@@ -37,7 +38,7 @@ bool TravelAction::Execute(Event /*event*/)
         if (!newTarget->IsAlive())
             continue;
 
-        if (newTarget->GetEntry() == target->getDestination()->getEntry())
+        if (newTarget->GetEntry() == uint32(target->getDestination()->getEntry()))
             continue;
 
         if (newTarget->IsInCombat())
